feat: implement 5.2 and 5.3
