+++ import-service/package.json
@@ -11,6 +11,7 @@
   "type": "module",
   "dependencies": {
     "@aws-sdk/client-s3": "^3.431.0",
+    "@aws-sdk/s3-request-presigner": "^3.435.0",
     "csv-parser": "^3.0.0"
   }
 }
